# API/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GNN Vulnerability Detection API",
    "description": "Inference (code -> CWE + vulnerable lines) and continual-learning relearn.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/inference": {
      "post": {
        "tags": [
          "inference"
        ],
        "summary": "Inference",
        "operationId": "inference_inference_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/embed": {
      "post": {
        "tags": [
          "inference"
        ],
        "summary": "Embed",
        "description": "Return the final pre-head function embedding per input \u2014 no classification output.\nFor similarity search and drift detection. Runs the same CPG + model pass as\n/inference but yields the representation fed to the output head, not the prediction.",
        "operationId": "embed_embed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inference/history": {
      "get": {
        "tags": [
          "inference"
        ],
        "summary": "Inference History",
        "description": "Recent persisted predictions (newest first).",
        "operationId": "inference_history_inference_history_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Inference History Inference History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/relearn": {
      "get": {
        "tags": [
          "relearn"
        ],
        "summary": "Relearn List",
        "operationId": "relearn_list_relearn_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Relearn List Relearn Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "relearn"
        ],
        "summary": "Relearn",
        "description": "Submit a continual-learning job (async).\n\nMaterializes the task-B dataset(s), trains a new model via the gnn_vuln library, evaluates\nit on the task-B test split (metrics-only), and registers the result. The library runs in\nAPI mode (`GNN_VULN_API_MODE`): no research artifacts \u2014 per-sample CSVs, ROC/confusion/PR\nplots, training log/curves \u2014 touch the worker disk. The metrics, the realized\ntrain/val/test split, and the training summary are persisted as per-model `model_artifacts`\n(DB + object storage), then the local results dir is deleted. Poll `GET /relearn/{job_id}`.",
        "operationId": "relearn_relearn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelearnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelearnJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/relearn/{job_id}": {
      "get": {
        "tags": [
          "relearn"
        ],
        "summary": "Relearn Status",
        "operationId": "relearn_status_relearn__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelearnJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/train": {
      "get": {
        "tags": [
          "train"
        ],
        "summary": "Train List",
        "operationId": "train_list_train_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Train List Train Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "train"
        ],
        "summary": "Train",
        "description": "Train a NEW model from scratch (async).\n\nArchitecture + train hyperparameters come from `config_id`; data from `dataset_ids`. Runs on\nthe same worker pipeline as /relearn but with fresh weights \u2014 no EWC\nimportance, no replay buffer. Registers the trained model and stores its metrics + split +\ntraining summary as model_artifacts. Poll `GET /train/{job_id}`. To CONTINUE an existing\nmodel instead of starting fresh, use `/relearn`.",
        "operationId": "train_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/train/{job_id}": {
      "get": {
        "tags": [
          "train"
        ],
        "summary": "Train Status",
        "operationId": "train_status_train__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmbedRequest": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "description": "Registered model id whose embedding space to use"
          },
          "codes": {
            "items": {
              "type": "string",
              "maxLength": 200000
            },
            "type": "array",
            "maxItems": 64,
            "minItems": 1,
            "title": "Codes",
            "description": "List of C/C++/Java function source strings"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "codes"
        ],
        "title": "EmbedRequest"
      },
      "EmbedResponse": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/EmbedResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "results"
        ],
        "title": "EmbedResponse"
      },
      "EmbedResult": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cls_embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cls Embedding",
            "description": "Final pre-head function representation (the vector fed to the output head). For similarity search / drift detection."
          },
          "cls_embedding_dim": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cls Embedding Dim"
          }
        },
        "type": "object",
        "required": [
          "index",
          "ok"
        ],
        "title": "EmbedResult"
      },
...
      "FunctionResult": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cached": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached"
          },
          "prediction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction"
          },
          "class_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Id"
          },
          "is_vulnerable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Vulnerable"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "class_probabilities": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Probabilities"
          },
          "suspicious_lines": {
            "items": {
              "$ref": "#/components/schemas/SuspiciousLine"
            },
            "type": "array",
            "title": "Suspicious Lines",
            "default": []
          },
          "cls_embedding_dim": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cls Embedding Dim",
            "description": "Dimensionality of the stored pre-head function embedding (full vector persisted for drift detection / similarity search, not returned in this response)"
          }
        },
        "type": "object",
        "required": [
          "index",
          "ok"
        ],
        "title": "FunctionResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InferenceConfig": {
        "properties": {
          "top_k_lines": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K Lines",
            "description": "Return only the top-k suspicious lines per function (all if null)"
          }
        },
        "type": "object",
        "title": "InferenceConfig",
        "description": "Inference-time options, kept separate from the call essentials (model_id + codes)."
      },
      "InferenceRequest": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "description": "Registered model id, e.g. 'graph_based', 'hybrid_graph_lm', 'sequential'"
          },
          "codes": {
            "items": {
              "type": "string",
              "maxLength": 200000
            },
            "type": "array",
            "maxItems": 64,
            "minItems": 1,
            "title": "Codes",
            "description": "List of C/C++/Java function source strings"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferenceConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Inference options (top_k_lines). Omit = defaults."
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "codes"
        ],
        "title": "InferenceRequest"
      },
      "InferenceResponse": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/FunctionResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "results"
        ],
        "title": "InferenceResponse"
      },
      "RelearnJob": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "dataset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dataset Ids"
          },
          "base_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Model Id"
          },
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "log_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Path"
          },
          "result_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Model Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics",
            "description": "Compact eval summary on the task-B test split (function-level classification + localization). The full metrics_summary, the training_summary, and the realized train/val/test split are stored as model_artifacts (kinds: metrics, training_summary, train_split)."
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "method",
          "dataset_ids"
        ],
        "title": "RelearnJob"
      },
      "RelearnMethod": {
        "type": "string",
        "enum": [
          "ER",
          "EWC",
          "EWC-ER",
          "finetune"
        ],
        "title": "RelearnMethod"
      },
      "RelearnRequest": {
        "properties": {
          "method": {
            "$ref": "#/components/schemas/RelearnMethod"
          },
          "dataset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dataset Ids",
            "description": "Task-B dataset id(s). More than one are joined before training."
          },
          "base_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Model Id",
            "description": "Required \u2014 the model to continue from. Its dataset, config and checkpoint are read from the registry. To train a fresh model instead, use POST /train."
          },
          "run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Name",
            "description": "Optional human label for the job"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Config overrides on the base model's config for this run (epochs, split). Everything else here is job spec, not config. Omit = use the base config as-is."
          }
        },
        "type": "object",
        "required": [
          "method",
          "dataset_ids"
        ],
        "title": "RelearnRequest"
      },
      "RunConfig": {
        "properties": {
          "epochs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epochs",
            "description": "Override training epochs"
          },
          "split": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SplitSpec"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional split control. Mode A: explicit {train,val,test} parquet_id lists. Mode B: {train_ratio,val_ratio,seed} for the library to split. Omit = default 80/10/10 seed 42."
          }
        },
        "type": "object",
        "title": "RunConfig",
        "description": "Config overrides applied on top of the inherited config for ONE run \u2014 kept separate from\nthe job spec (method, datasets, base_model). Omit any field to keep the inherited value."
      },
...
      "TrainJob": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "dataset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dataset Ids"
          },
          "config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Id"
          },
          "config_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Path"
          },
          "log_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Path"
          },
          "result_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Model Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics",
            "description": "Compact eval summary on the test split (function-level classification + localization). Full artifacts are stored as model_artifacts, same as /relearn."
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "dataset_ids"
        ],
        "title": "TrainJob"
      },
      "TrainRequest": {
        "properties": {
          "config_id": {
            "type": "string",
            "title": "Config Id",
            "description": "Registered config to instantiate the model (architecture + train hyperparameters). The dataset(s) provide the data."
          },
          "dataset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Dataset Ids",
            "description": "Dataset id(s) to train on. More than one are joined first."
          },
          "run_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Name",
            "description": "Optional human label for the job"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Config overrides on config_id for this run (epochs, split). Omit = use config_id as-is."
          }
        },
        "type": "object",
        "required": [
          "config_id",
          "dataset_ids"
        ],
        "title": "TrainRequest",
        "description": "Train a brand-new model from scratch. The architecture + train hyperparameters come from\n`config_id`; the data from `dataset_ids`. (Continuing an existing model is `/relearn`.)"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}